bug fix after merging PeriodSelectivity

+++ pipeline/foraging_model.py
@@ -442,7 +442,7 @@ def get_session_history(session_key, remove_ignored=True):
     reward_history = np.zeros([2, len(_reward)])  # .shape = (2, N trials)
     for c in (0, 1):
         reward_history[c, _choice == c] = (_reward[_choice == c] == True).astype(int)
-    choice_history = np.array([_choice]).astype(int)  # .shape = (1, N trials)
+    choice_history = np.array([_choice]).astype(float)  # .shape = (1, N trials)
     
     # -- ITI --
     # All previous models has an effective ITI of constant 1.
